requirements.txt
nsw-fuel-api-client==1.1.0
influxdb-client>=1.40.0
PyYAML>=6.0
schedule>=1.2.0
python-dotenv>=1.0.0
